Refactor monetary columns to use BigDecimal and update database schema (#724)
* Refactor monetary columns to use BigDecimal and update database schema

- Changed NAV and transaction-related fields from Float/Double to BigDecimal in entities and DTOs.
- Updated mapping logic to handle BigDecimal for NAV values.
- Modified application properties for improved timeout settings.
- Created Liquibase migration scripts to alter database columns to NUMERIC type for monetary values.
- Aligned timestamp columns to use timestamptz in PostgreSQL.
- Added foreign key indexes for improved query performance.
- Updated version columns to SMALLINT and adjusted sequences for user_cas_details and user_folio_details.
- Implemented trigram search indexes for enhanced search capabilities on scheme names.

* Fix migration data integrity and align NAV constraints and search indexes

Preserve unit precision, deduplicate NAV rows, require UTC confirmation, reseed sequences, and add migration integration coverage.

---------

+++ src/main/java/com/learning/mfscreener/models/NAVDataDTO.java
@@ -3,10 +3,13 @@ package com.learning.mfscreener.models;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record NAVDataDTO(
-        @JsonFormat(pattern = "dd-MM-yyyy", shape = JsonFormat.Shape.STRING) LocalDate date, Float nav, Long schemeId)
+        @JsonFormat(pattern = "dd-MM-yyyy", shape = JsonFormat.Shape.STRING) LocalDate date,
+        BigDecimal nav,
+        Long schemeId)
         implements Serializable {
 
     public NAVDataDTO withSchemeId(Long schemeCode) {
